property changes maintenance

+++ app/Http/Controllers/Property/PropertyMaintenanceRequestController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Property;
 
+use App\Enums\Core\PermissionEnum;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Auth;
@@ -22,41 +23,38 @@ class PropertyMaintenanceRequestController extends Controller
         $maintenancerequests = PropertyMaintenanceRequest::all();
         return view('property.maintenanceandissues.maintenancerequest.index', compact('maintenancerequests'));
     }
-
     public function create(){
-        //$this->authorize(PermissionEnum::PropertyMaintenanceRequestUpdate, PropertyMaintenanceRequest::class);
-        $properties = PropertyRegistry::all();
+        $this->authorize(PermissionEnum::PropertyMaintenanceRequestUpdate, PropertyMaintenanceRequest::class);
+        $properties = PropertyRegistry::all()->where('IsActive',True);
         $issuetypes = CodeDetail::where('CodeID', 'IssueType')->get();
         $priorities = CodeDetail::where('CodeID', 'PriorityLevel')->get();
         return view('property.maintenanceandissues.maintenancerequest.create', compact('properties','issuetypes','priorities'));
     }
-        public function show($Id){
-       // $this->authorize(PermissionEnum::PropertyMaintenanceRequestView, PropertyMaintenanceRequest::class);
+    public function show($Id){
+        $this->authorize(PermissionEnum::PropertyMaintenanceRequestView, PropertyMaintenanceRequest::class);
         $maintenancerequest = PropertyMaintenanceRequest::with('issueType','priority')->get()->find($Id);
         return view('property.maintenanceandissues.maintenancerequest.show',compact('maintenancerequest'));
     }
-    public function getBlockByProperty($propertyId)
+    public function getBlockByProperty($PropertyId)
     {
-        $blocks = PropertyBlock::where('PropertyID', $propertyId)->get();
+        $blocks = PropertyBlock::where('PropertyID', $PropertyId)->get();
         return response()->json($blocks);
     }
-
-
-    public function getFloorByBlock($blockId)
+    public function getFloorByBlock($BlockId)
     {
-        $floors = PropertyFloor::where('BlockID', $blockId)->get();
+        $floors = PropertyFloor::where('BlockID', $BlockId)->get();
         return response()->json($floors);
     }
-    public function getUnitByFloor($floorId)
+    public function getUnitByFloor($FloorId)
     {
-        $units = PropertyUnit::where('FloorId', $floorId)->get();
+        $units = PropertyUnit::where('FloorId', $FloorId)->get();
         return response()->json($units);
     }
 
     public function store(MaintenanceRequest $request)
     {
-       // $this->authorize(PermissionEnum::PropertyMaintenanceRequestCreate, PropertyMaintenanceRequest::class);
-        //dd($request->all());
+        $this->authorize(PermissionEnum::PropertyMaintenanceRequestCreate, PropertyMaintenanceRequest::class);
+
         $validated = $request->validated();
 
         $Property = PropertyRegistry::findOrFail($validated['Property']);
@@ -86,7 +84,7 @@ class PropertyMaintenanceRequestController extends Controller
     }
     public function edit($Id)
     {
-      //  $this->authorize(PermissionEnum::PropertyMaintenanceRequestUpdate, PropertyMaintenanceRequest::class);
+        $this->authorize(PermissionEnum::PropertyMaintenanceRequestUpdate, PropertyMaintenanceRequest::class);
         $maintenancerequest = PropertyMaintenanceRequest::findOrFail($Id);
         $properties = PropertyRegistry::all();
         $issuetypes = CodeDetail::where('CodeID', 'IssueType')->get();
@@ -96,8 +94,8 @@ class PropertyMaintenanceRequestController extends Controller
 
     public function update(MaintenanceRequest $request, $Id)
     {
-      //  $this->authorize(PermissionEnum::PropertyMaintenanceRequestUpdate, PropertyMaintenanceRequest::class);
-        // Validate the request data
+        $this->authorize(PermissionEnum::PropertyMaintenanceRequestUpdate, PropertyMaintenanceRequest::class);
+ 
         $validated = $request->validated();
         DB::beginTransaction();
 
@@ -142,7 +140,7 @@ class PropertyMaintenanceRequestController extends Controller
 
             DB::commit();
             return redirect()->route('maintenancerequest.index')
-                            ->with('success', 'Maintenance request updated successfully');
+            ->with('success', 'Maintenance request updated successfully');
         } catch (\Throwable $th) {
             DB::rollBack();
             Log::error('Failed to update maintenance request: ' . $th->getMessage());
@@ -153,7 +151,7 @@ class PropertyMaintenanceRequestController extends Controller
 
     public function destroy($Id)
     {
-       // $this->authorize(PermissionEnum::PropertyMaintenanceRequestDelete, PropertyMaintenanceRequest::class);
+        $this->authorize(PermissionEnum::PropertyMaintenanceRequestDelete, PropertyMaintenanceRequest::class);
         try {
             $maintenancerequest = PropertyMaintenanceRequest::findOrFail($Id);
             $maintenancerequest->delete();
